* behoben: Datenbankmigrationen brechen nicht mehr ab, da in seed.rb das commerce_data Modul, und darin das CSV-Modul geladen werden

+++ db/app_migrations/034_create_frontend_errors.rb
@@ -1,24 +1,21 @@
 class CreateFrontendErrors < Sequel::Migration
-    def up
-      create_table :frontend_errors do
-        primary_key :id
-        String :user_agent
-        String :platform
-        String :type
-        String :frontend_version
-        String :backend_version
-        String :message
-        String :stacktrace
-        String :ip_hash
+  def up
+    create_table :frontend_errors do
+      primary_key :id
+      String :user_agent
+      String :platform
+      String :type
+      String :frontend_version
+      String :backend_version
+      String :message
+      String :stacktrace
+      String :ip_hash
 
-        DateTime :created_at
-      end
-    end
-  
-  
-    def down
-      if @db.table_exists?(:frontend_errors)
-        drop_table :frontend_errors
-      end
+      DateTime :created_at
     end
   end
+
+  def down
+    drop_table :frontend_errors if @db.table_exists?(:frontend_errors)
+  end
+end
